# 伤害行为新增随机数值波动 # 伤害频率流派 - 意外惊喜: 友军触发伤害时会有随机物理伤害加成〔2%-20%攻击力〕

+++ Assets/Script/Bussiness/Logic/Action/Main/GameActionValueFormat.cs
@@ -14,7 +14,7 @@ namespace GamePlay.Bussiness.Logic
 
     public static class GameActionValueFormatUtil
     {
-        public static float FormatValue(this GameActionValueFormat format, int value)
+        public static float FormatValue(this GameActionValueFormat format, float value)
         {
             var formatV = 0f;
             switch (format)
